ui/requirements.txt
numpy==1.15.0
PyQt5==5.11.2
PyQt5-sip==4.19.12
pyqtgraph==0.10.0
pyzmq==17.1.0
